Improved unit test for results

+++ voting/voting-vote/voting-vote-api/src/main/java/com/prodyna/voting/vote/VotingOptionResult.java
@@ -11,5 +11,5 @@ public class VotingOptionResult {
 
     private int countVotes;
 
-    private double percentage;
+    private float percentage;
 }
